Converted detail-card tests. Adjusting process so that git history is kept on files
