+++ karma.conf.js
@@ -11,7 +11,7 @@ module.exports = config => {
 				//
 				// npm run test -- --grep test/foo/bar.test.js
 				// npm run test -- --grep test/bar/*
-				{ pattern: config.grep ? config.grep : 'components/**/*.test.js', type: 'module' },
+				{ pattern: config.grep ? config.grep : 'test/**/*.test.js', type: 'module' },
 			],
 			// see the karma-esm docs for all options
 			esm: {
